Add roles

+++ TuristickaAgencija/TuristickaAgencija/Controllers/RezervacijasController.cs
@@ -9,10 +9,12 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using TuristickaAgencija.Data;
 using TuristickaAgencija.Models;
+using TuristickaAgencija.Strings;
 
 namespace TuristickaAgencija.Controllers
 {
-    [Authorize]
+    [Authorize(Roles =RoleNames.AdminIkorisnik)]
+    
     public class RezervacijasController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -25,12 +27,18 @@ namespace TuristickaAgencija.Controllers
         // GET: Rezervacijas
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.Rezervacije
-                .Include(r => r.Prevoz)
-                .Include(r => r.Smestaj)
-                .ThenInclude(s=>s.Aranzman)
-                .Include(r => r.User);
-            return View(await applicationDbContext.ToListAsync());
+            List<Rezervacija> rezervacijas;
+
+            if (User.IsInRole(RoleNames.Admin))
+            {
+                rezervacijas = await GetRezervacijas();
+            }
+            else
+            {
+                string currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+                rezervacijas = await GetRezervacijasByUserId(currentUserId);
+            }
+            return View(rezervacijas);
         }
 
         // GET: Rezervacijas/Details/5
@@ -56,6 +64,7 @@ namespace TuristickaAgencija.Controllers
         }
 
         // GET: Rezervacijas/Create
+        [Authorize(Roles=RoleNames.Korisnik)]
         public IActionResult Create(Guid smestajId)
         {
             var smestaj = _context
@@ -80,6 +89,7 @@ namespace TuristickaAgencija.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles=RoleNames.Korisnik)]
         public async Task<IActionResult> Create([Bind("UserId,SmestajId,PrevozId,Id")] Rezervacija rezervacija)
         {
             if (ModelState.IsValid)
@@ -110,13 +120,13 @@ namespace TuristickaAgencija.Controllers
 
 
         // GET: Rezervacijas/Delete/5
+        [Authorize(Roles = RoleNames.Korisnik)]
         public async Task<IActionResult> Delete(Guid? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
-
             var rezervacija = await _context.Rezervacije
                 .Include(r => r.Prevoz)
                 .Include(r => r.Smestaj)
@@ -128,15 +138,29 @@ namespace TuristickaAgencija.Controllers
                 return NotFound();
             }
 
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (rezervacija.UserId != currentUserId)
+            {
+                return Forbid();
+            }
+
             return View(rezervacija);
         }
 
         // POST: Rezervacijas/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles=RoleNames.Korisnik)]
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
             var rezervacija = await _context.Rezervacije.FindAsync(id);
+            if(rezervacija.UserId != currentUserId)
+            {
+                return Forbid();
+            }
             _context.Rezervacije.Remove(rezervacija);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -145,6 +169,27 @@ namespace TuristickaAgencija.Controllers
         private bool RezervacijaExists(Guid id)
         {
             return _context.Rezervacije.Any(e => e.Id == id);
+        }
+
+        private Task<List<Rezervacija>> GetRezervacijas()
+        {
+            var rezervacijas =  _context.Rezervacije
+               .Include(r => r.Prevoz)
+               .Include(r => r.Smestaj)
+               .ThenInclude(s => s.Aranzman)
+               .Include(r => r.User);
+            return rezervacijas.ToListAsync();
+        }
+
+        private Task<List<Rezervacija>> GetRezervacijasByUserId(string userId)
+        {
+            var rezervacijas = _context.Rezervacije
+               .Include(r => r.Prevoz)
+               .Include(r => r.Smestaj)
+               .ThenInclude(s => s.Aranzman)
+               .Include(r => r.User)
+               .Where(r => r.UserId == userId);
+            return rezervacijas.ToListAsync();
         }
     }
 }
